Integrate Metadata move in Infrastructure layer
The services.yml needed to be updated. The LoadMetadataCommandHandler
also required some updates.

Finally the FakeFetcher now also implements the correct interface.
