+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Resources/config/services.yml
@@ -32,6 +32,10 @@ services:
 
     surfnet.dashboard.command_handler.load_metadata:
         class: Surfnet\ServiceProviderDashboard\Application\CommandHandler\Service\LoadMetadataCommandHandler
+        arguments:
+            - '@surfnet.dashboard.repository.service'
+            - '@surfnet.dashboard.metadata.fetcher'
+            - '@surfnet.dashboard.metadata.parser'
         public: true
         tags:
             - { name: tactician.handler, command: Surfnet\ServiceProviderDashboard\Application\Command\Service\LoadMetadataCommand }
@@ -46,29 +50,29 @@ services:
         class: Surfnet\ServiceProviderDashboard\Application\Factory\ServiceCommandFactory
 
     surfnet.dashboard.metadata.certificate_parser:
-        class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Metadata\CertificateParser
+        class: Surfnet\ServiceProviderDashboard\Legacy\Metadata\CertificateParser
 
     surfnet.dashboard.metadata.client:
         class: GuzzleHttp\Client
 
     surfnet.dashboard.metadata.fetcher:
-        class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Metadata\Fetcher
+        class: Surfnet\ServiceProviderDashboard\Legacy\Metadata\Fetcher
         arguments: ['@surfnet.dashboard.metadata.client', '@logger', '%metadata_url_timeout%']
 
     surfnet.dashboard.metadata.parser:
-        class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Metadata\Parser
+        class: Surfnet\ServiceProviderDashboard\Legacy\Metadata\Parser
         arguments:
             - '@surfnet.dashboard.metadata.certificate_parser'
             - '@surfnet.dashboard.repository.attributes_metadata'
-            - '%kernel.project_dir%/src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Resources'
+            - '%kernel.root_dir%/Resources'
             - '@logger'
 
     surfnet.dashboard.monolog.json_formatter:
         class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Monolog\Formatter\JsonFormatter
 
     surfnet.dashboard.repository.attributes_metadata:
-        class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Repository\AttributesMetadataRepository
-        arguments:  ['%kernel.project_dir%/src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Resources']
+        class: Surfnet\ServiceProviderDashboard\Legacy\Repository\AttributesMetadataRepository
+        arguments:  ['%kernel.root_dir%/Resources']
 
     surfnet.dashboard.repository.service:
         class: Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Repository\ServiceRepository
